+++ lib/src/interactive/scene_controller_scene.dart
@@ -4,70 +4,75 @@ import '../contract/node_patch.dart';
 import '../contract/node_spec.dart';
 import '../contract/scene_write_txn.dart';
 import '../contract/snapshot.dart';
-import 'internal/scene_controller_scene_access.dart';
+import 'internal/scene_controller_scene_mutations.dart';
 
 class SceneControllerScene {
-  SceneControllerScene(this._access);
+  const SceneControllerScene({
+    required this.ensurePublicSideEffectAllowed,
+    required SceneControllerSceneMutations mutations,
+  }) : _mutations = mutations;
 
-  final SceneControllerSceneAccess _access;
+  final void Function(String operation, {bool allowAfterDispose})
+  ensurePublicSideEffectAllowed;
+  final SceneControllerSceneMutations _mutations;
 
   T write<T>(T Function(SceneWriteTxn writer) fn) {
-    _access.ensurePublicSideEffectAllowed('write');
-    return _access.write(fn);
+    ensurePublicSideEffectAllowed('write');
+    return _mutations.write(fn);
   }
 
   void setBackgroundColor(Color value) {
-    _access.ensurePublicSideEffectAllowed('setBackgroundColor');
-    _access.setBackgroundColor(value);
+    ensurePublicSideEffectAllowed('setBackgroundColor');
+    _mutations.setBackgroundColor(value);
   }
 
   void setGridEnabled(bool value) {
-    _access.ensurePublicSideEffectAllowed('setGridEnabled');
-    _access.setGridEnabled(value);
+    ensurePublicSideEffectAllowed('setGridEnabled');
+    _mutations.setGridEnabled(value);
   }
 
   void setGridCellSize(double value) {
-    _access.ensurePublicSideEffectAllowed('setGridCellSize');
-    _access.setGridCellSize(value);
+    ensurePublicSideEffectAllowed('setGridCellSize');
+    _mutations.setGridCellSize(value);
   }
 
   void setCameraOffset(Offset value) {
-    _access.ensurePublicSideEffectAllowed('setCameraOffset');
-    _access.setCameraOffset(value);
+    ensurePublicSideEffectAllowed('setCameraOffset');
+    _mutations.setCameraOffset(value);
   }
 
   NodeId addNode(NodeSpec node, {LayerId? layerId, int? insertIndex}) {
-    _access.ensurePublicSideEffectAllowed('addNode');
-    return _access.addNode(node, layerId: layerId, insertIndex: insertIndex);
+    ensurePublicSideEffectAllowed('addNode');
+    return _mutations.addNode(node, layerId: layerId, insertIndex: insertIndex);
   }
 
   bool ensureLayer(LayerId layerId, {int? index}) {
-    _access.ensurePublicSideEffectAllowed('ensureLayer');
-    return _access.ensureLayer(layerId, index: index);
+    ensurePublicSideEffectAllowed('ensureLayer');
+    return _mutations.ensureLayer(layerId, index: index);
   }
 
   bool patchNode(NodePatch patch) {
-    _access.ensurePublicSideEffectAllowed('patchNode');
-    return _access.patchNode(patch);
+    ensurePublicSideEffectAllowed('patchNode');
+    return _mutations.patchNode(patch);
   }
 
   bool removeNode(NodeId id, {int? timestampMs}) {
-    _access.ensurePublicSideEffectAllowed('removeNode');
-    return _access.removeNode(id, timestampMs: timestampMs);
+    ensurePublicSideEffectAllowed('removeNode');
+    return _mutations.removeNode(id, timestampMs: timestampMs);
   }
 
   void clearScene({int? timestampMs}) {
-    _access.ensurePublicSideEffectAllowed('clearScene');
-    _access.clearScene(timestampMs: timestampMs);
+    ensurePublicSideEffectAllowed('clearScene');
+    _mutations.clearScene(timestampMs: timestampMs);
   }
 
   void replaceScene(SceneSnapshot snapshot) {
-    _access.ensurePublicSideEffectAllowed('replaceScene');
-    _access.replaceScene(snapshot);
+    ensurePublicSideEffectAllowed('replaceScene');
+    _mutations.replaceScene(snapshot);
   }
 
   void notifySceneChanged() {
-    _access.ensurePublicSideEffectAllowed('notifySceneChanged');
-    _access.notifySceneChanged();
+    ensurePublicSideEffectAllowed('notifySceneChanged');
+    _mutations.notifySceneChanged();
   }
 }