+++ lib/src/view/scene_view_interactive.dart
@@ -123,6 +123,7 @@ class _SceneViewInteractiveState extends State<SceneViewInteractive> {
           _pointerHost.handlePointerEvent(event, PointerPhase.cancel),
       child: CustomPaint(
         foregroundPainter: SceneViewInteractiveOverlayPainter(
+          controller: widget.controller,
           interaction: widget.controller.interaction,
         ),
         child: renderSurface,